integrated new Image and ImageDS classes into belpy code
