+++ lib/base_classes/images.py
@@ -101,7 +101,7 @@ def process_resize(w, h, resize):
 class Image:
     def __init__(self, path: Union[str, Path], image: np.ndarray = None) -> None:
 
-        self._path = path
+        self._path = Path(path)
         self._value_array = None
         self._width = None
         self._height = None
@@ -141,29 +141,43 @@ class Image:
 
     @property
     def date(self):
-        return self._date_time
+        if self._date_time is not None:
+            return self._date_time.strftime("%Y:%m:%d")
+        else:
+            print("No exif data available.")
 
-    def get_image(self) -> np.ndarray:
-        """Returns the image"""
+    @property
+    def time(self):
+        if self._date_time is not None:
+            return self._date_time.strftime("%H:%M:%S")
+        else:
+            print("No exif data available.")
+
+    @property
+    def value(self) -> np.ndarray:
+        """Returns the image as numpy array"""
         if self._value_array is not None:
             return self._value_array
         else:
             return self.read_image(self._path)
 
+    def get_datetime(self):
+        return self._date_time
+
     def read_image(
         self,
-        path: Union[str, Path],
+        # path: Union[str, Path],
         col: bool = True,
         resize: List[int] = [-1],
         crop: List[int] = None,
     ) -> None:
         """Wrapper around the function read_image to be a class method."""
-        path = Path(path)
-        if path.exists():
-            self._value_array = read_image(path, col, resize, crop)
+        # path = Path(path)
+        if self.path.exists():
+            self._value_array = read_image(self.path, col, resize, crop)
             self.read_exif()
         else:
-            print(f"Input paht {path} not valid.")
+            print(f"Input paht {self.path} not valid.")
 
     def clean_image(self) -> None:
         self._value_array = None
@@ -177,7 +191,7 @@ class Image:
         except:
             print("No exif data available.")
 
-        # Set image size
+        # Get image size
         if (
             "Image ImageWidth" in self._exif_data.keys()
             and "Image ImageLength" in self._exif_data.keys()
@@ -185,17 +199,22 @@ class Image:
             self._width = self._exif_data["Image ImageWidth"].printable
             self._height = self._exif_data["Image ImageLength"].printable
 
-        # Set Image Datetime
+        # Get Image Date and Time
+        self._date_time_fmt = "%Y:%m:%d %H:%M:%S"
         if "Image DateTime" in self._exif_data.keys():
-            date_fmt = "%Y:%m:%d %H:%M:%S"
             date_str = self._exif_data["Image DateTime"].printable
-            self._date_time = datetime.strptime(date_str, date_fmt)
+        elif "EXIF DateTimeOriginal" in self._exif_data.keys():
+            date_str = self._exif_data["EXIF DateTimeOriginal"].printable
+        else:
+            print("Date not available in exif.")
+            return
+        self._date_time = datetime.strptime(date_str, self._date_time_fmt)
 
-    def extract_patch(self, limits: dict) -> np.ndarray:
+    def extract_patch(self, limits: List[int]) -> np.ndarray:
         """Extract image patch
         Parameters
         __________
-        - limits (dict): dictionary containing the index of the tile (in row-major order, C-style) and a list of the bounding box coordinates as: {0,[xmin, xmax, ymin, ymax]}
+        - limits (List[int]): List containing the bounding box coordinates as: [xmin, xmax, ymin, ymax]
         __________
         Return: patch (np.ndarray)
         """
@@ -219,102 +238,149 @@ class Image:
         Returns:
             intrinsics matrix (3x3).
         """
-
         # if self._exif_data is None or len(self._exif_data) == 0:
         #     return None
-
         # focal_length_mm = self.exif_data.get("FocalLength")
-
         # sensor_width_mm = Image.sensor_width_db.lookup(
         #     self._exif_data.get("Make"),
         #     self._exif_data.get("Model"),
         # )
-
         # img_w_px = self._width
         # img_h_px = self._height
         # focal_length_px = max(img_h_px, img_w_px) * \
         #     focal_length_mm / sensor_width_mm
-
         # center_x = img_w_px / 2
         # center_y = img_h_px / 2
 
 
-class Imageds:
+class ImageDS:
     """
-    Class to help manage Image datasets
+    Class to manage Image datasets for multi epoch
 
     """
 
     def __init__(
         self,
-        path=None,
+        folder: Union[str, Path],
+        ext: str = None,
+        recursive: bool = False,
         logger: logging = None,
-    ):
-        # TODO: implement labels in datastore
-        if not hasattr(self, "files"):
-            self.reset_imageds()
-        if path is not None:
-            self.get_image_list(path)
+    ) -> None:
+        self.reset_imageds()
 
-    def __len__(self):
+        self.folder = Path(folder)
+        if not self.folder.exists():
+            msg = "Error: invalid input path."
+            if logger is not None:
+                logger.error(msg)
+            else:
+                print(msg)
+            return
+        if ext is not None:
+            self.ext = ext
+        self.recursive = recursive
+
+        self.read_image_list(self.folder)
+
+    def __len__(self) -> int:
         """Get number of images in the datastore"""
         return len(self.files)
 
-    def __contains__(self, name):
+    def __contains__(self, name: str) -> bool:
         """Check if an image is in the datastore, given the image name"""
-        return name in self.files
+        files = [x.name for x in self.files]
+        return name in files
 
-    def __getitem__(self, idx, **args):
-        """Read and return the image at position idx in the image datastore"""
-        # TODO: add possibility to chose reading between col or grayscale, scale image, crop etc...
-        # @TODO change getitem to return path and implement reading function
-    
-        img = read_image(os.path.join(self.folder[idx], self.files[idx]))
-        assert img is not None, f"Unable to read image {self.files[idx]}"
-    
-        return img
+    def __getitem__(self, idx: int) -> str:
+        """Return image name (including extension) at position idx in datastore"""
 
-    def reset_imageds(self):
+        return self.files[idx].name
+
+    def read_image(self, idx: int) -> Image:
+        """Return image at position idx as Image instance, containing both exif and value data (accessible by value proprierty, e.g., image.value)"""
+        image = Image(self.files[idx])
+        image.read_image()
+        return image
+
+    def reset_imageds(self) -> None:
         """Initialize image datastore"""
-        self.files = []
-        self.folder = []
-        self.ext = []
-        self.label = []
-        # self.size = []
-        # self.shot_date = []
-        # self.shot_time = []
+        self.files = None
+        self.folder = None
+        self.ext = None
 
-    def get_image_list(self, path):
-        # TODO: change name in read image list
-        # TODO: add option for including subfolders
-        if not os.path.exists(path):
-            print("Error: invalid input path.")
-            return
-        d = os.listdir(path)
-        d.sort()
-        self.files = d
-        self.folder = [path] * len(d)
+    def read_image_list(self, recursive: bool = None) -> None:
+        assert self.folder.is_dir(), "Error: invalid image directory."
 
-    def get_image_name(self, idx):
-        """Return image name at position idx in datastore"""
+        if recursive is not None:
+            self.recursive = recursive
+        if self.recursive:
+            rec_patt = "**/"
+        else:
+            rec_patt = ""
+        if self.ext is not None:
+            ext_patt = f".{self.ext}"
+        else:
+            ext_patt = ""
+        pattern = f"{rec_patt}*{ext_patt}"
+
+        self.files = sorted(self.folder.glob(pattern))
+        if len(self.files) == 0:
+            print(f"No images found in folder {self.folder}")
+
+    def get_image_path(self, idx: int) -> Path:
+        """Return path of the image at position idx in datastore as Pathlib"""
         return self.files[idx]
 
-    def get_image_path(self, idx):
-        """Return full path of the image at position idx in datastore"""
-        return os.path.join(self.folder[idx], self.files[idx])
-
-    def get_image_stem(self, idx):
+    def get_image_stem(self, idx: int) -> str:
         """Return name without extension(stem) of the image at position idx in datastore"""
-        return Path(self.files[idx]).stem
+        return self.files[idx].stem
+
+    def write_exif_to_csv(
+        self, filename: str, sep: str = ",", header: bool = True
+    ) -> None:
+        assert self.folder.is_dir(), "Empty Image Datastore."
+        file = open(filename, "w")
+        if header:
+            file.write("epoch,name,date,time\n")
+        for i, img_path in enumerate(self.files):
+            img = Image(img_path)
+            name = img_path.name
+            date = img.date
+            time = img.time
+            file.write(f"{i}{sep}{name}{sep}{date}{sep}{time}\n")
+        file.close()
 
 
 if __name__ == "__main__":
     """Test classes"""
 
-    cams = ["p1", "p2"]
-    images = dict.fromkeys(cams)
-    for cam in cams:
-        images[cam] = Imageds(Path("data/img2022") / cam)
+    # images = ImageDS("data/img2022/p1")
 
-    im = Image(images["p1"].get_image_path(0))
-    print(im)
+    # # Get image name
+    # images[0]
+
+    # # Get image stem
+    # images.get_image_stem(0)
+
+    # # Get image path
+    # images.get_image_path(0)
+
+    # # Get image as Image object and extect date and time
+    # img = images.read_image(0)
+    # img.date
+    # img.time
+
+    # # Read image as numpy array
+    # image = images.read_image(0).value
+
+    # # Write exif to csv file
+    # filename = "test.csv"
+    # images.write_exif_to_csv(filename)
+
+    cams = ["p1", "p2"]
+    images = {}
+    for cam in cams:
+        images[cam] = ImageDS(Path("data/img2021") / cam)
+        images[cam].write_exif_to_csv(f"data/img2021/image_list_{cam}.csv")
+
+    print("Done")
