+++ lib/geometry.py
@@ -137,7 +137,7 @@ def undistort_image_new_cam_matrix(image, K, dist, downsample=1, out_path=None):
         cv2.imwrite(out_path, und)
     return und, K_scaled
     # cam = 1
-    # image = images[cam][0]
+    # image = images[cam].read_image(epoch).value
     # K, dist = cameras[cam][0].K, cameras[cam][0].dist
     # image_und = cv2.undistort(image, K, dist, None, K)
     # cv2.imwrite(images[cam].get_image_stem(0)+'_undistorted.tif', image_und)
