+++ lib/utils/inizialize_variables.py
@@ -29,7 +29,7 @@ from typing import List, Union
 
 from lib.base_classes.camera import Camera
 from lib.base_classes.pointCloud import PointCloud
-from lib.base_classes.images import Image, Imageds
+from lib.base_classes.images_new import Image, ImageDS
 from lib.base_classes.targets import Targets
 from lib.base_classes.features import Features
 
@@ -40,11 +40,11 @@ class Inizialization:
         self.cfg = cfg
         self.cams = self.cfg.paths.camera_names
 
-    def init_image_ds(self) -> List[Imageds]:
+    def init_image_ds(self) -> dict:
         # Create Image Datastore objects
         images = dict.fromkeys(self.cams)
         for cam in self.cams:
-            images[cam] = Imageds(self.cfg.paths.image_dir / cam)
+            images[cam] = ImageDS(self.cfg.paths.image_dir / cam)
 
         self.images = images
         return self.images
@@ -55,19 +55,20 @@ class Inizialization:
             image = Image(self.images[self.cams[0]].get_image_path(epoch))
             epoch_dict[epoch] = Path(
                 (self.cfg.paths.results_dir)
-                / f"{image.date.year}_{image.date.month:02}_{image.date.day:02}"
+                / f"{image.get_datetime().year}_{image.get_datetime().month:02}_{image.get_datetime().day:02}"
             ).stem
 
         self.epoch_dict = epoch_dict
         return self.epoch_dict
 
-    def init_cameras(self) -> List[edict]:
+    def init_cameras(self) -> dict:
         cameras = {}
 
         assert hasattr(
             self, "images"
         ), "Images datastore not available yet. Inizialize images first"
-        im_height, im_width = self.images[self.cams[0]][0].shape[:2]
+        img = Image(self.images[self.cams[0]].get_image_path(0))
+        im_height, im_width = img.height, img.width
 
         # Inizialize Camera Intrinsics at every epoch setting them equal to the those of the reference cameras.
         for epoch in self.cfg.proc.epoch_to_process:
